Create weather_app
This is my first time learning about and trying to create my own app that utilizes an API. This was constructed with the help of realpython.com, without whom the struggle would have been immense.

The main problem I have been unable to yet solve is adding color and style to the output. The big success for me is the fact that my API requests actually worked. The goal when I started this project was to understand the basic parts of an API request and I feel like its a real win that the function works like I intended it to.